refactor: 프록시 쿠키 재작성 로직을 공용 유틸로 분리
applyProxyAuthResponse.ts와 proxy/[...path]/route.ts에 toProxyCookiePath 등
동일한 쿠키 파싱/재작성 로직이 중복돼 있어, oauth_state 스코프 변경 같은
작업 시 두 파일을 매번 같이 고쳐야 했음. shared/lib/proxyCookie.ts로 추출해
drift 위험을 제거.
